Subdivide traffic models into proper package hierarchy
