+++ src/main/kotlin/tinder/gold/adventures/chronos/listener/TrackSensorListener.kt
@@ -5,7 +5,7 @@ import org.eclipse.paho.client.mqttv3.MqttAsyncClient
 import org.eclipse.paho.client.mqttv3.MqttMessage
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
-import tinder.gold.adventures.chronos.model.traffic.control.TrainTrack
+import tinder.gold.adventures.chronos.model.traffic.track.TrainTrack
 import tinder.gold.adventures.chronos.mqtt.getPayloadString
 import tinder.gold.adventures.chronos.service.TrafficFilterService
