+++ src/main/kotlin/tinder/gold/adventures/chronos/model/mqtt/builder/MqttTopicBuilderSubject.kt
@@ -2,7 +2,7 @@ package tinder.gold.adventures.chronos.model.mqtt.builder
 
 import tinder.gold.adventures.chronos.model.mqtt.builder.MqttTopicBuilder.CardinalDirection
 import tinder.gold.adventures.chronos.model.mqtt.builder.MqttTopicBuilder.CardinalDirection.*
-import tinder.gold.adventures.chronos.model.traffic.control.ITrafficControl
+import tinder.gold.adventures.chronos.model.traffic.core.ITrafficControl
 
 open class MqttTopicBuilderSubject(
         val LANE_TYPE: MqttTopicBuilder.LaneType,
