+++ src/main/kotlin/tinder/gold/adventures/chronos/model/traffic/barrier/VesselControlBarrier.kt
@@ -1,6 +1,7 @@
-package tinder.gold.adventures.chronos.model.traffic.control
+package tinder.gold.adventures.chronos.model.traffic.barrier
 
 import tinder.gold.adventures.chronos.model.mqtt.builder.MqttTopicBuilder
+import tinder.gold.adventures.chronos.model.traffic.core.TrafficControlBarrier
 
 class VesselControlBarrier(
         componentId: Int
