+++ src/main/kotlin/tinder/gold/adventures/chronos/model/traffic/core/IControlBarrier.kt
@@ -1,4 +1,4 @@
-package tinder.gold.adventures.chronos.model.traffic.control
+package tinder.gold.adventures.chronos.model.traffic.core
 
 import org.eclipse.paho.client.mqttv3.MqttAsyncClient
 import tinder.gold.adventures.chronos.model.mqtt.MqttPublisher
