+++ src/main/kotlin/tinder/gold/adventures/chronos/model/traffic/core/WarningLight.kt
@@ -1,4 +1,4 @@
-package tinder.gold.adventures.chronos.model.traffic.control
+package tinder.gold.adventures.chronos.model.traffic.core
 
 abstract class WarningLight(
         override val componentId: Int = 0
