+++ src/main/kotlin/tinder/gold/adventures/chronos/model/traffic/light/CycleTrafficLight.kt
@@ -1,20 +1,20 @@
-package tinder.gold.adventures.chronos.model.traffic.control
+package tinder.gold.adventures.chronos.model.traffic.light
 
 import tinder.gold.adventures.chronos.model.mqtt.MqttPublisher
 import tinder.gold.adventures.chronos.model.mqtt.MqttSubscriber
 import tinder.gold.adventures.chronos.model.mqtt.builder.MqttTopicBuilder
 import tinder.gold.adventures.chronos.model.mqtt.builder.MqttTopicBuilder.CardinalDirection
+import tinder.gold.adventures.chronos.model.traffic.core.TrafficLight
 
-class MotorisedTrafficLight(
+class CycleTrafficLight(
         directionTo: CardinalDirection,
-        overrideSubgroupId: Int? =null
+        overrideSubgroupId: Int? = null
 ) : TrafficLight(
         directionTo,
         MqttTopicBuilder.LaneType.MOTORISED,
         MqttTopicBuilder.ComponentType.TRAFFIC_LIGHT,
         overrideSubgroupId
 ) {
-
     override lateinit var publisher: MqttPublisher
     override lateinit var subscriber: MqttSubscriber
 }
