+++ src/main/kotlin/tinder/gold/adventures/chronos/model/traffic/light/TrafficLightState.kt
@@ -1,10 +1,9 @@
-package tinder.gold.adventures.chronos.model.traffic.control
+package tinder.gold.adventures.chronos.model.traffic.light
 
 import org.eclipse.paho.client.mqttv3.MqttAsyncClient
 import tinder.gold.adventures.chronos.model.mqtt.MqttPublisher
 
 sealed class TrafficLightState {
-
     object Green : TrafficLightState() {
         override fun getPayload() = "2"
     }
