+++ src/main/kotlin/tinder/gold/adventures/chronos/model/traffic/light/VesselWarningLight.kt
@@ -1,8 +1,10 @@
-package tinder.gold.adventures.chronos.model.traffic.control
+package tinder.gold.adventures.chronos.model.traffic.light
 
 import tinder.gold.adventures.chronos.model.mqtt.MqttPublisher
 import tinder.gold.adventures.chronos.model.mqtt.MqttSubscriber
 import tinder.gold.adventures.chronos.model.mqtt.builder.MqttTopicBuilder
+import tinder.gold.adventures.chronos.model.traffic.core.IWarningLight
+import tinder.gold.adventures.chronos.model.traffic.core.WarningLight
 
 class VesselWarningLight : WarningLight() {
     override val overrideSubgroup: Int? = null
