+++ src/main/kotlin/tinder/gold/adventures/chronos/service/TrafficControlService.kt
@@ -8,7 +8,7 @@ import mu.KotlinLogging
 import org.eclipse.paho.client.mqttv3.MqttAsyncClient
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import tinder.gold.adventures.chronos.model.traffic.control.TrafficLight
+import tinder.gold.adventures.chronos.model.traffic.core.TrafficLight
 import tinder.gold.adventures.chronos.model.traffic.sensor.TrafficSensor
 import javax.annotation.PostConstruct
